chef ability somewhat buggy destroy behaviour

+++ unity-ktwo/Assets/Scripts/Behaviours/BasePuddleBehaviour.cs
@@ -6,7 +6,7 @@ public class BasePuddleBehaviour : MonoBehaviour
 {
     [Header("Puddle Settings")]
     public int numberOfUses;
-    public int decayRate;
+    public float decayRate;
 
     public HashSet<GameObject> affectedEntities = new HashSet<GameObject>();
 
@@ -14,6 +14,7 @@ public class BasePuddleBehaviour : MonoBehaviour
     {
         StartCoroutine(PuddleLifetime(decayRate));
     }
+
     void Update()
     {
         if (numberOfUses <= 0 && affectedEntities.Count == 0) Destroy(gameObject);
